fix changes in handler and separete db and app
