+++ models/boil_table_names.go
@@ -5,6 +5,8 @@ package models
 
 var TableNames = struct {
 	Universities string
+	Users        string
 }{
 	Universities: "universities",
+	Users:        "users",
 }
